Implement EPH (Exception Prefix High) bit.
The significant bits (31-12) of the vector offset address for each
exception depend on the setting of the Supervision Register (SR)'s EPH
bit and setting of of the Exception Vector Base Address Register
(EVBAR), which can specify an offset.

In the absence of the EVBAR and with SR[EPH] clear, the offset is zero.

If SR[EPH] is set, this value is logically ORed with the offset
0xF0000000.

This means if EPH is;
 * 0 - Exceptions vectors start at (EVBAR << 8)
 * 1 - Exception vectors start at (EVBAR << 8) | 0xF0000000

Signed-off-by: Tim 'mithro' Ansell <mithro@mithis.com>